requirements.txt
selenium>=4.15.0
webdriver-manager>=4.0.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
